remove original looper cli
